% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebtf.R
\name{ebtf}
\alias{ebtf}
\title{Fit empirical Bayes trend filtering/dynamic linear model via variational inference.}
\usage{
ebtf(
  x,
  v,
  ord = 0,
  sk = NULL,
  w_init = NULL,
  m_init = NULL,
  maxiter = 500,
  tol = 1e-05,
  printevery = Inf,
  calc_obj_every = 1,
  plot_every = Inf,
  lambda0 = 100
)
}
\arguments{
\item{x, v}{data vector and variance}

\item{ord}{the order of trend filtering, currently can be 0,1, or 2.}

\item{sk}{grid of prior standard deviations}

\item{w_init}{init of prior weights}

\item{m_init}{init of smoothed curve/posterior mean}

\item{maxiter, tol}{criteria to stop the iteration}

\item{printevery}{how often to print results}

\item{calc_obj_every}{how often to calculate elbo}

\item{plot_every}{how often to plot results}

\item{lambda0}{The parameter promoting further sparsity/smoothness}
}
\value{
A list of: posterior, fitted_g
}
\description{
Assuming the posterior factorizes as q(b,z) = q(b)q(z)
}
\examples{
library(ebtf)
set.seed(123)
n=2^9
xx = DJ.EX(n,rsnr = 3)
mu = xx$blocks
v = rep(1,n)
v[sample(1:n,round(n*0.1))] = 10
v[sample(1:n,round(n*0.1))] = 0.1
x = mu + rnorm(n,0,sqrt(v))

fit = ebtf(x,v)
plot(x,col='grey80')
lines(mu,col='grey50')
lines(fit$posterior$mean)
}
